% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rill-package.R
\docType{package}
\name{rill-package}
\alias{rill}
\alias{rill-package}
\title{rill: A personal feed reader}
\description{
rill is a small, personal feed reader built with Shiny. It keeps feed
ingestion, clean reading copies, durable state, and operational telemetry
behind a focused application interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jameshwade.github.io/rill/}
  \item \url{https://github.com/JamesHWade/rill}
  \item Report bugs at \url{https://github.com/JamesHWade/rill/issues}
}

}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com}

Authors:
\itemize{
  \item James Wade \email{github@jameshwade.com}
}

}
\keyword{internal}
